Changed lengths and positions from int64_t to size_t.

+++ src/keys/rsakey.cpp
@@ -38,7 +38,7 @@ CryptoPP::RSAFunction ReadRSAPublicKey(std::string key_material) {
     throw parse4880::invalid_header_error(-1);
   }
 
-  int64_t modulus_length = parse4880::ReadInteger(key_material.substr(0,2));
+  size_t modulus_length = parse4880::ReadInteger(key_material.substr(0,2));
 
   modulus_length = ((modulus_length+7) / 8);
   if (key_material.length() < 2 + modulus_length) {
@@ -49,7 +49,7 @@ CryptoPP::RSAFunction ReadRSAPublicKey(std::string key_material) {
   modulus.OpenPGPDecode(reinterpret_cast<const uint8_t*>(
       key_material.substr(0, 2+modulus_length).c_str()), 2+modulus_length);
 
-  int64_t exponent_length = parse4880::ReadInteger(
+  size_t exponent_length = parse4880::ReadInteger(
       key_material.substr(2+modulus_length,2));
   exponent_length = ((exponent_length +7) / 8);
   if (key_material.length() < 4 + modulus_length + exponent_length) {
